refactor: move posts to features folder
