+++ src/components/features/user/UserAccountNav.tsx
@@ -1,16 +1,16 @@
-"use client";
+'use client';
 
-import React, { FC } from "react";
+import React, { FC } from 'react';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "./DropDownMenu";
-import UserAvatar from "./UserAvatar";
-import { signOut } from "next-auth/react";
-import { MoreVertical } from "lucide-react";
-import { User } from "@prisma/client";
+} from '../../ui/DropDownMenu';
+import UserAvatar from './UserAvatar';
+import { signOut } from 'next-auth/react';
+import { MoreVertical } from 'lucide-react';
+import { User } from '@prisma/client';
 
 interface UserAccountNavProps {
   user: User | null;
@@ -21,10 +21,10 @@ const UserAccountNav: FC<UserAccountNavProps> = ({ user }) => {
   const displayName = user?.displayName ?? user?.name;
 
   return (
-    <div className="space-y-2">
-      <div className="flex items-center justify-between">
+    <div className='space-y-2'>
+      <div className='flex items-center justify-between'>
         <UserAvatar
-          className="h-11 w-11"
+          className='h-11 w-11'
           user={{
             name: zUser?.name || null,
             image: zUser?.image || null,
@@ -33,7 +33,7 @@ const UserAccountNav: FC<UserAccountNavProps> = ({ user }) => {
 
         <DropdownMenu>
           <DropdownMenuTrigger>
-            <MoreVertical className="h-4 w-4" />
+            <MoreVertical className='h-4 w-4' />
           </DropdownMenuTrigger>
           <DropdownMenuContent>
             <DropdownMenuItem
@@ -43,7 +43,7 @@ const UserAccountNav: FC<UserAccountNavProps> = ({ user }) => {
                   callbackUrl: `${window.location.origin}/sign-in`,
                 });
               }}
-              className="cursor-pointer"
+              className='cursor-pointer'
             >
               Sign Out
             </DropdownMenuItem>
@@ -51,10 +51,10 @@ const UserAccountNav: FC<UserAccountNavProps> = ({ user }) => {
         </DropdownMenu>
       </div>
 
-      <div className="flex flex-col space-y-1 leading-none">
-        {displayName && <p className="font-medium">{displayName}</p>}
+      <div className='flex flex-col space-y-1 leading-none'>
+        {displayName && <p className='font-medium'>{displayName}</p>}
         {zUser?.username && (
-          <p className="w-[200px] truncate text-sm text-muted-foreground">
+          <p className='w-[200px] truncate text-sm text-muted-foreground'>
             {`u/${zUser.username}`}
           </p>
         )}
