+++ src/components/feeds/following/FollowingFeed.tsx
@@ -6,7 +6,7 @@ import { useIntersection } from '@mantine/hooks';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { useSession } from 'next-auth/react';
-import Post from '../../posts/Post';
+import Post from '../../features/posts/Post';
 import { Loader2 } from 'lucide-react';
 import { INFINITE_SCROLLING_PAGINATION_RESULTS } from '@/config';
 import { Suspense } from 'react';
